% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duluthEditablemap.R
\name{duluthEditablemap}
\alias{duluthEditablemap}
\title{deluthEditablemap data}
\format{
## `duluthEditablemap`
A data frame with 1946 rows and 4 columns:
\describe{
  \item{fid}{ID}
  \item{geometry}{geometry}
  ...
}
}
\usage{
data(duluthEditablemap)
}
\description{
A polygon map
Report ...
}
