% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duluthBasemap.R
\name{duluthBasemap}
\alias{duluthBasemap}
\title{deluthBasemap data}
\format{
## `duluthBasemap`
A data frame with 45 rows and 5 columns:
\describe{
  \item{TRACTFIPS}{ID}
  \item{NRI_FEMA}{NRI_FEMA}
  \item{SVI_CDC}{SVI_CDC}
  \item{geometry}{geometry}
  ...
}
}
\usage{
data(duluthBasemap)
}
\description{
A polygon map
Report ...
}
