% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duluthEditablemap.R
\name{duluthEditablemap}
\alias{duluthEditablemap}
\title{deluthEditablemap data}
\format{
## `duluthEditablemap`
A data frame with 1946 rows and 4 columns:
\describe{
  \item{fid}{GEOID}
  \item{geometry}{geometry}
  ...
}
}
\source{
<https://data-duluthmn.opendata.arcgis.com/search?tags=Basemap>
}
\usage{
data(duluthEditablemap)
}
\description{
A polygon map of the city of Duluth, Minnesota. The data include social and environmental vulnerabilities for all census tracts of the city
}
