% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  editable_map = PPGISr::duluthEditablemap \%>\% dplyr::mutate(PPGIS_CODE =
    as.character(dplyr::row_number()), SELECTED = NA) \%>\% dplyr::select(PPGIS_CODE,
    SELECTED, geometry) \%>\% sf::st_transform(4326),
  base_map = PPGISr::duluthBasemap,
  mapping_categories = c("High Density Development", "Street Trees",
    "Infrastructure Need"),
  mapping_colors = c("#880015", "#22b14c", "#00a2e8"),
  editable_map_info_icon_message =
    "Click the circle to left to choose mapping categories you want to add to the map. Click the map to indicate these preferences",
  basemap_info_icon_message =
    "This allows you to view data for making your mapping decisions",
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{...}{arguments to pass to golem_opts.
See `?golem::get_golem_options` for more details.}
}
\description{
Run the Shiny Application
}
\examples{
\dontrun{

run_app()

run_app(mapping_categories = c("flowers", "lake", "trees"))

}


}
