% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{A function to run the PPGISr Shiny Application}
\usage{
run_app(
  editable_map = PPGISr::duluthEditablemap,
  base_map = PPGISr::duluthBasemap,
  basemap_name = "Reference Basemap",
  mapping_categories = c("High Density Development", "Street Trees",
    "Infrastructure Need"),
  mapping_colors = c("#880015", "#22b14c", "#00a2e8"),
  editable_map_info_icon_message =
    "Click the circle to left to choose mapping categories you want to add to the map. Click the map to indicate these preferences",
  basemap_info_icon_message =
    "This allows you to view data for making your mapping decisions",
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{editable_map}{A vector file (e.g., multiploygon, polygon)}

\item{base_map}{A vector (e.g., multipolygon, polygon) or raster (i.e. tif, grid) file for visualization}

\item{basemap_name}{The legend name given to your imported base map}

\item{mapping_categories}{Vector of names for mapping categories options (i.e. "Trees", "Flowers","Water")}

\item{mapping_colors}{Vector of colors corresponding to vector of names (i.e. "green", "red","blue")}

\item{editable_map_info_icon_message}{Message for your editable info button}

\item{basemap_info_icon_message}{Message for basemap info button}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{...}{arguments to pass to golem_opts.
See `?golem::get_golem_options` for more details.}
}
\description{
A function to run the PPGISr Shiny Application
}
\examples{
\dontrun{

PPGISr::run_app()
## Customize your PPGISr interface by adding mapping categories
run_app(mapping_categories = c("flowers", "lake", "trees"))

## Customize your PPGISr interface by specifying colors for your mapping categories
## This must be a vector of the same length as the mapping categories
run_app(mapping_categories = c("flowers", "lake", "trees"),
        mapping_colors = c("green", "red", "blue"))

## Customize your PPGISr interface by specifying an editable map
## We recommend using the sf library to load spatial vector data
library(sf)
## Spatial data should use World Geodetic System 1984 (WGS84) coordinate reference system(crs)
editable_map <- st_read(system.file("shape/nc.shp", package="sf"))

run_app(editable_map = editable_map, base_map = NULL)
}


}
